+++ config-plugins/checkboxes-plugin.js
@@ -4,7 +4,7 @@ const checkboxes = plugin(function ({ addUtilities, theme }) {
     ".checkbox_style": {
       height: "44px",
       width: "44px",
-      borderColor: theme("colors.multi.neutrals.grey85a"),
+      // borderColor: theme("colors.multi.neutrals.grey85a"),
       borderRadius: "4px",
       borderWidth: "1.5px",
     },
@@ -13,7 +13,6 @@ const checkboxes = plugin(function ({ addUtilities, theme }) {
       opacity: "1",
       "&:focus": {
         "--tw-ring-color": theme("colors.multi.blue.blue60f"),
-        backgroundColor: theme("colors.specific.green.green70"),
       },
     },
   };