+++ tailwind.config.js
@@ -1,6 +1,7 @@
 const plugin = require("tailwindcss/plugin");
 
 module.exports = {
+  prefix: "ds-",
   content: [
     "./src/**/*.{js,ts,jsx,tsx}",
     "./config-plugins/**/*.{js,ts,jsx,tsx}",
@@ -169,7 +170,11 @@ module.exports = {
         "color-grid": "repeat(auto-fill, minmax(180px, 1fr))",
       },
       spacing: {
+        "22px": "22px",
         "30px": "30px",
+        "42px": "42px",
+        "38px": "38px",
+        "18px": "18px",
       },
       width: {
         "40px": "40px",
@@ -181,6 +186,7 @@ module.exports = {
         "138px": "138px",
         "160px": "160px",
         "168px": "168px",
+        "284px": "284px",
         "300px": "300px",
         "400px": "400px",
         "500px": "500px",
@@ -249,6 +255,7 @@ module.exports = {
         "600px": "600px",
         "750px": "750px",
         "5rem": "5rem",
+        "250px": "250px",
       },
 
       minWidth: {
@@ -264,6 +271,7 @@ module.exports = {
         "200px": "200px",
         "250px": "250px",
         "288px": "288px",
+        "44px": "44px",
       },
 
       maxHeight: {
@@ -284,6 +292,7 @@ module.exports = {
         "33px": "33px",
         "22px": "22px",
         "20px": "20px",
+        "24px": "24px",
       },
       outline: {
         "black-solid": ["2px solid #000", "1px"],
@@ -318,5 +327,6 @@ module.exports = {
     require("./config-plugins/cards-plugin"),
     require("./config-plugins/labels-plugin"),
     require("./config-plugins/checkboxes-plugin"),
+    require("./config-plugins/search-plugin"),
   ],
 };